Fix issue loading a single string array resource
The problem was an out of bounds error while executing the
update statement in the for loop to extract the contents of the string
array. Switch to a while loop to avoid executing the update statement unless
it is guaranteed that there is an additional item.

Test will be forthcoming but may require refactoring the testapp directory.
A test will require a separate app that contains a single string
array resource file.

PiperOrigin-RevId: 315504525